Refine runtime narrative and selection pipeline

+++ C/t1_v2/minute_ring.h
@@ -5,7 +5,7 @@
 namespace t1_v2 {
 
 struct MinuteSlot {
-    int minute_index = -1;
+    int64_t minute_index = -1;
     int px_milli = 0;
     int64_t amt_yuan = 0;
     // Shares, not lots. Must match RawTick::vol_units.
@@ -15,7 +15,7 @@ struct MinuteSlot {
 struct MinuteRingState {
     static constexpr int KEEP = 8;
     MinuteSlot slots[KEEP];
-    int latest_minute = -1;
+    int64_t latest_minute = -1;
 };
 
 }  // namespace t1_v2
